// src/app/modules/project-setup/common-setup/models/onstreet/parking.model.ts
export class Parking {
  id: string;
  number: number;
  parking_code: number;
  name: string;
  latitude: number;
  longitude: number;
  parking_angle: number;
  parking_length: number;
  parking_spaces: number;
  parking_dimension: string;
  is_sensors: boolean;
  parking_type: string;
  managed_by: string;
  project_id: number;
  zone_id: string;
  spaces_nbr_from: number;
  spaces_nbr_to: number;
  pictures_url: string;
  info_notes: string;
  connecting_points: string;
  name_ar: string;
  functioning: string;
  restriction: string;
  payment_methods: number[];
}